+++ internal/ent/mixin/audit_mixin.go
@@ -1,4 +1,4 @@
-package schema
+package mixin
 
 import (
 	"context"
