+++ internal/ent/privacy/rule/rule.go
@@ -11,7 +11,7 @@ import (
 	"github.com/datumforge/datum/internal/ent/generated"
 	"github.com/datumforge/datum/internal/ent/generated/privacy"
 	"github.com/datumforge/datum/internal/ent/generated/user"
-	"github.com/datumforge/datum/privacy/viewer"
+	"github.com/datumforge/datum/internal/ent/privacy/viewer"
 )
 
 // DenyIfNoViewer is a rule that returns deny decision if the viewer is missing in the context.
