+++ internal/ent/schema/session.go
@@ -9,6 +9,7 @@ import (
 	"entgo.io/ent/schema/edge"
 	"entgo.io/ent/schema/field"
 	"entgo.io/ent/schema/index"
+	"github.com/datumforge/datum/internal/ent/mixin"
 	"github.com/google/uuid"
 	"lukechampine.com/frand"
 )
@@ -99,6 +100,6 @@ func (Session) Annotations() []schema.Annotation {
 // Mixin of the Session
 func (Session) Mixin() []ent.Mixin {
 	return []ent.Mixin{
-		AuditMixin{},
+		mixin.AuditMixin{},
 	}
 }
